# docs/swagger.yaml
basePath: /
definitions:
  domain.CreateTroveRequest:
    properties:
      description:
        type: string
      title:
        type: string
    type: object
  domain.CreateUserRequest:
    properties:
      avatar:
        example: https://example.com/avatar.jpg
        type: string
      email:
        example: john@example.com
        type: string
      nickname:
        example: John Doe
        type: string
      password:
        example: "123456"
        minLength: 6
        type: string
      status:
        example: 1
        type: integer
      username:
        example: john_doe
        type: string
    required:
    - email
    - password
    - username
    type: object
...
  domain.UpdateTroveRequest:
    properties:
      description:
        type: string
      title:
        type: string
    type: object
  domain.UpdateUserRequest:
    properties:
      avatar:
        example: https://example.com/avatar.jpg
        type: string
      email:
        example: john@example.com
        type: string
      nickname:
        example: John Doe
        type: string
      password:
        example: "123456"
        minLength: 6
        type: string
      status:
        example: 1
        type: integer
      username:
        example: john_doe
        type: string
    type: object
...
  model.Trove:
    description: 收藏信息
    properties:
      created_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      description:
        type: string
      id:
        example: 1
        type: integer
      title:
        type: string
      updated_at:
        example: "2023-01-01T00:00:00Z"
        type: string
    type: object
  model.User:
    description: 用户信息
    properties:
      avatar:
        example: https://example.com/avatar.jpg
        type: string
      created_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      email:
        example: john@example.com
        type: string
      id:
        example: 1
        type: integer
      nickname:
        example: John Doe
        type: string
      status:
        description: '1: 正常, 0: 禁用'
        example: 1
        type: integer
      updated_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      username:
        example: john_doe
        type: string
    type: object
...
  pagination.PageResponse:
    properties:
      data:
        description: 数据列表
      has_next:
        description: 是否有下一页
        type: boolean
      has_prev:
        description: 是否有上一页
        type: boolean
      page:
        description: 当前页码
        type: integer
      page_size:
        description: 每页大小
        type: integer
      total:
        description: 总记录数
        type: integer
      total_pages:
        description: 总页数
        type: integer
    type: object
  response.Response:
    properties:
      code:
        description: 状态码
        type: integer
      data:
        description: 响应数据
      error:
        description: 错误信息
        type: string
      message:
        description: 响应消息
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a dove server API documentation.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: dove API
  version: "1.0"
paths:
...
  /troves/:
    get:
      consumes:
      - application/json
      description: 获取所有Trove列表（支持分页、排序和搜索）
      parameters:
      - description: 页码，默认为1
        in: query
        minimum: 1
        name: page
        type: integer
      - description: 每页大小，默认为10，最大100
        in: query
        maximum: 100
        minimum: 1
        name: page_size
        type: integer
      - description: '排序字段：ID, '
        in: query
        name: sort_by
        type: string
      - description: 排序方向：asc, desc，默认为desc
        in: query
        name: sort_order
        type: string
      - description: 搜索关键词
        in: query
        name: keyword
        type: string
      - description: 搜索字段：Title, Description, ，不指定则在所有字段中搜索
        in: query
        name: search_by
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/pagination.PageResponse'
                  - properties:
                      data:
                        items:
                          $ref: '#/definitions/model.Trove'
                        type: array
                    type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: 获取所有Trove
      tags:
      - troves
    post:
      consumes:
      - application/json
      description: 创建新的Trove
      parameters:
      - description: Trove信息
        in: body
        name: trove
        required: true
        schema:
          $ref: '#/definitions/domain.CreateTroveRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Trove'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: 创建Trove
      tags:
      - troves
  /troves/{id}:
    delete:
      consumes:
      - application/json
      description: 根据ID删除Trove
      parameters:
      - description: TroveID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: 删除Trove
      tags:
      - troves
    get:
      consumes:
      - application/json
      description: 根据ID获取Trove详细信息
      parameters:
      - description: TroveID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Trove'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      summary: 获取Trove详情
      tags:
      - troves
    put:
      consumes:
      - application/json
      description: 根据ID更新Trove信息
      parameters:
      - description: TroveID
        in: path
        name: id
        required: true
        type: integer
      - description: Trove更新信息
        in: body
        name: trove
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateTroveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Trove'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: 更新Trove
      tags:
      - troves
  /users:
    post:
      consumes:
      - application/json
      description: 创建新用户
      parameters:
      - description: 用户信息
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/domain.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.User'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: 创建用户
      tags:
      - 用户
  /users/:
    get:
      consumes:
      - application/json
      description: 获取所有用户列表（支持分页、排序和搜索）
      parameters:
      - description: 页码，默认为1
        in: query
        minimum: 1
        name: page
        type: integer
      - description: 每页大小，默认为10，最大100
        in: query
        maximum: 100
        minimum: 1
        name: page_size
        type: integer
      - description: 排序字段：id, username, email, nickname, status, created_at, updated_at
        in: query
        name: sort_by
        type: string
      - description: 排序方向：asc, desc，默认为desc
        in: query
        name: sort_order
        type: string
      - description: 搜索关键词
        in: query
        name: keyword
        type: string
      - description: 搜索字段：username, email, nickname，不指定则在所有字段中搜索
        in: query
        name: search_by
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/pagination.PageResponse'
                  - properties:
                      data:
                        items:
                          $ref: '#/definitions/model.User'
                        type: array
                    type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: 获取所有用户
      tags:
      - 用户
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: 根据ID删除用户
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: 删除用户
      tags:
      - 用户
    get:
      consumes:
      - application/json
      description: 根据ID获取用户详细信息
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.User'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      summary: 获取用户详情
      tags:
      - 用户
    put:
      consumes:
      - application/json
      description: 根据ID更新用户信息
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      - description: 用户更新信息
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.User'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: 更新用户
      tags:
      - 用户
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"